+++ promotion/src/main/java/com/klp/promotion/coupon/domain/repository/UserCouponRepository.java
@@ -1,6 +1,7 @@
 package com.klp.promotion.coupon.domain.repository;
 
 import com.klp.promotion.coupon.domain.entity.UserCoupon;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,4 +16,6 @@ public interface UserCouponRepository {
     List<UserCoupon> findAllByUserId(Long userId);
 
     int reserve(UUID userCouponId, Integer version);
+
+    void markExpiredCoupons(LocalDateTime todayStart);
 }
