+++ promotion/src/main/java/com/klp/promotion/coupon/infrastructure/repository/UserCouponRepositoryImpl.java
@@ -1,8 +1,13 @@
 package com.klp.promotion.coupon.infrastructure.repository;
 
+import com.klp.promotion.coupon.domain.entity.QCoupon;
+import com.klp.promotion.coupon.domain.entity.QUserCoupon;
 import com.klp.promotion.coupon.domain.entity.UserCoupon;
 import com.klp.promotion.coupon.domain.enums.UserCouponStatus;
 import com.klp.promotion.coupon.domain.repository.UserCouponRepository;
+import com.querydsl.jpa.JPAExpressions;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +18,7 @@ import org.springframework.stereotype.Repository;
 public class UserCouponRepositoryImpl implements UserCouponRepository {
 
     private final UserCouponJpaRepotiory userCouponJpaRepotiory;
+    private final JPAQueryFactory queryFactory;
 
     @Override
     public UserCoupon findByUserIdAndCouponId(Long userId, UUID couponId) {
@@ -36,5 +42,25 @@ public class UserCouponRepositoryImpl implements UserCouponRepository {
     @Override
     public int reserve(UUID userCouponId, Integer version) {
         return userCouponJpaRepotiory.reserve(userCouponId, version);
+    }
+
+    @Override
+    public void markExpiredCoupons(LocalDateTime todayStart) {
+        QUserCoupon userCoupon = QUserCoupon.userCoupon;
+        QCoupon coupon = QCoupon.coupon;
+
+        queryFactory
+            .update(userCoupon)
+            .set(userCoupon.status, UserCouponStatus.EXPIRED)
+            .where(
+                userCoupon.status.eq(UserCouponStatus.READY),
+                userCoupon.couponId.in(
+                    JPAExpressions
+                        .select(coupon.couponId)
+                        .from(coupon)
+                        .where(coupon.expired_at.lt(todayStart))
+                )
+            )
+            .execute();
     }
 }
